+++ src/components/Calculators/Electronics/Basic/StandardResistance/MainView.vue
@@ -60,14 +60,11 @@
 
 <script>
 
-//  'use strict'
-
   import Calc from 'src/misc/CalculatorEngineV2/Calc'
   import {CalcVarNumeric} from 'src/misc/CalculatorEngineV2/CalcVarNumeric'
   import {UnitMulti} from 'src/misc/CalculatorEngineV2/UnitMulti'
   import PresetValidators from 'src/misc/CalculatorEngineV2/PresetValidators'
   import { CustomValidator } from 'src/misc/CalculatorEngineV2/CustomValidator'
-
   import StandardResistanceFinder from 'src/misc/StandardResistanceFinder/StandardResistanceFinder'
   import ESeriesRow from './ESeriesRow.vue'
 
@@ -120,18 +117,13 @@
         ],
         helpText: 'The resistance you actually want. The closest value to this resistance will be found in each resistor series.'
       }))
-
-      // Configure calculator to default state now that all
-      // variables have been added.
       calc.init()
-
       return {
         calc: calc,
         standardResistanceFinder: standardResistanceFinder
       }
     },
     mounted () {
-//      console.log('Standard Resistance calculator mounted.')
       window.MathJax.Hub.Queue(['Typeset', window.MathJax.Hub])
     }
   }