+++ src/main.js
@@ -1,7 +1,7 @@
 // The Vue build version to load with the `import` command
 // (runtime-only or standalone) has been set in webpack.base.conf with an alias.
 import Vue from 'vue'
-import App from './components/App'
+import App from './components/App/App'
 
 /* eslint-disable no-unused-vars */
 
@@ -29,6 +29,8 @@ Vue.use(VTooltip)
 import ElementUI from 'element-ui'
 import 'element-ui/lib/theme-default/index.css'
 Vue.use(ElementUI)
+
+import './style/style.css'
 
 // =========================================== //
 // ==== CALCULATOR COMPONENT REGISTRATION ==== //